requirements.txt
Django==3.2.8
djangorestframework==3.12.4
django-cors-headers==3.10.0
djangorestframework-jwt==1.11.0
PyJWT==2.3.0
